Build a parent class
