+++ servlet/src/main/java/com/springRaft/servlet/consensusModule/Follower.java
@@ -1,6 +1,8 @@
 package com.springRaft.servlet.consensusModule;
 
 import com.springRaft.servlet.communication.message.*;
+import com.springRaft.servlet.communication.outbound.OutboundManager;
+import com.springRaft.servlet.config.RaftProperties;
 import com.springRaft.servlet.persistence.state.StateService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,22 +14,10 @@ import java.util.concurrent.ScheduledFuture;
 
 @Service
 @Scope("singleton")
-public class Follower implements RaftState {
+public class Follower extends RaftStateContext implements RaftState {
 
     /* Logger */
     private static final Logger log = LoggerFactory.getLogger(Follower.class);
-
-    /* Application Context for getting beans */
-    private final ApplicationContext applicationContext;
-
-    /* Module that has the consensus functions to invoke */
-    private final ConsensusModule consensusModule;
-
-    /* Service to access persisted state repository */
-    private final StateService stateService;
-
-    /* Transition handler */
-    private final TransitionManager transitionManager;
 
     /* Scheduled Thread */
     private ScheduledFuture<?> scheduledFuture;
@@ -38,12 +28,15 @@ public class Follower implements RaftState {
             ApplicationContext applicationContext,
             ConsensusModule consensusModule,
             StateService stateService,
-            TransitionManager transitionManager
+            RaftProperties raftProperties,
+            TransitionManager transitionManager,
+            OutboundManager outboundManager
     ) {
-        this.applicationContext = applicationContext;
-        this.consensusModule = consensusModule;
-        this.stateService = stateService;
-        this.transitionManager = transitionManager;
+        super(
+                applicationContext, consensusModule,
+                stateService, raftProperties,
+                transitionManager, outboundManager
+        );
         this.scheduledFuture = null;
     }
 
@@ -63,29 +56,7 @@ public class Follower implements RaftState {
     @Override
     public AppendEntriesReply appendEntries(AppendEntries appendEntries) {
 
-        AppendEntriesReply reply = this.applicationContext.getBean(AppendEntriesReply.class);
-
-        long currentTerm = this.stateService.getCurrentTerm();
-
-        if (appendEntries.getTerm() < currentTerm) {
-
-            reply.setTerm(currentTerm);
-            reply.setSuccess(false);
-
-        } else if (appendEntries.getTerm() > currentTerm) {
-
-            // update term
-            this.stateService.setState(appendEntries.getTerm(), null);
-
-            this.setAppendEntriesReply(appendEntries, reply);
-
-        } else if (appendEntries.getTerm() == currentTerm) {
-
-            this.setAppendEntriesReply(appendEntries, reply);
-
-        }
-
-        return reply;
+        return super.appendEntries(appendEntries);
 
     }
 
@@ -175,79 +146,13 @@ public class Follower implements RaftState {
     /* --------------------------------------------------- */
 
     /**
-     * Set a timer in milliseconds that represents a timeout.
-     * */
-    private void setTimeout() {
-
-        // schedule task
-        ScheduledFuture<?> schedule = this.transitionManager.setElectionTimeout();
-
-        // store runnable
-        this.setScheduledFuture(schedule);
-
-    }
-
-    /**
-     * TODO
-     * */
-    private void checkLog(RequestVote requestVote, RequestVoteReply reply) {
-
-        if (requestVote.getLastLogTerm() > this.consensusModule.getCommittedTerm()) {
-
-            // vote for this request if not voted for anyone yet
-            this.setVote(requestVote, reply);
-
-        } else if (requestVote.getLastLogTerm() < this.consensusModule.getCommittedTerm()) {
-
-            // revoke request
-            reply.setVoteGranted(false);
-
-        } else if (requestVote.getLastLogTerm() == this.consensusModule.getCommittedTerm()) {
-
-            if (requestVote.getLastLogIndex() >= this.consensusModule.getCommittedIndex()) {
-
-                // vote for this request if not voted for anyone yet
-                this.setVote(requestVote, reply);
-
-            } else if (requestVote.getLastLogIndex() < this.consensusModule.getCommittedIndex()) {
-
-                // revoke request
-                reply.setVoteGranted(false);
-
-            }
-
-        }
-
-    }
-
-    /**
-     * TODO
-     * */
-    private void setVote(RequestVote requestVote, RequestVoteReply reply) {
-
-        String votedFor = this.stateService.getVotedFor();
-
-        if (votedFor == null || votedFor.equals(requestVote.getCandidateId())) {
-
-            this.stateService.setVotedFor(requestVote.getCandidateId());
-            reply.setVoteGranted(true);
-
-        } else {
-
-            reply.setVoteGranted(false);
-
-        }
-
-    }
-
-    /**
      * A method that encapsulates replicated code, and has the function of setting
      * the reply for the received AppendEntries.
      *
      * @param appendEntries The received AppendEntries communication.
      * @param reply AppendEntriesReply object, to send as response to the leader.
      * */
-    private void setAppendEntriesReply(AppendEntries appendEntries, AppendEntriesReply reply) {
+    protected void setAppendEntriesReply(AppendEntries appendEntries, AppendEntriesReply reply) {
 
         // remove the scheduled task
         this.transitionManager.cancelScheduledTask(this.scheduledFuture);
@@ -264,6 +169,21 @@ public class Follower implements RaftState {
 
         // set a new timeout, it's equivalent to transit to a new follower state
         this.setTimeout();
+
+    }
+
+    /* --------------------------------------------------- */
+
+    /**
+     * Set a timer in milliseconds that represents a timeout.
+     * */
+    private void setTimeout() {
+
+        // schedule task
+        ScheduledFuture<?> schedule = this.transitionManager.setElectionTimeout();
+
+        // store runnable
+        this.setScheduledFuture(schedule);
 
     }
